+++ Rtos5.h
@@ -10,6 +10,14 @@
 #define TIME_DELAY_FIRST_FIX 10 //10 navpvt messages alvorens start loggen
 #define TIME_DELAY_NEW_RUN 10 //int time_delay_new_run
 #define EPOCH_2022 1640995200 //start of the year 2022 1640995200
+#define UBLOX_TYPE_UNKNOWN 0
+#define M8_9600BD 1
+#define M10_9600BD 2
+#define M8_38400BD 3
+#define M10_38400BD 4
+#define M9_9600BD 5
+#define M9_38400BD 6
+#define AUTO_DETECT 0xFF
 
 #if defined(STATIC_DEBUG)
 #define MIN_SPEED_START_LOGGING 0